+++ packages/core/src/index.ts
@@ -11,6 +11,7 @@ export {
 export { handleCodeTitles } from "./plugins/handleCodeTitles";
 export { handleCodeExpandableRemark, handleCodeExpandable } from "./plugins/handleCodeExpandable";
 export { rehypeMermaid } from "./plugins/rehypeMermaid";
+export { remarkDirectiveToMdx } from "./plugins/remarkDirectiveToMdx";
 export {
   extractFrontmatter,
   extractFrontmatterWithContent,