+++ packages/flame/vitest.config.ts
@@ -9,20 +9,20 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      "@docubook/ui-react/input": path.resolve(__dirname, "../ui/react/dist/input.js"),
-      "@docubook/ui-react/kbd": path.resolve(__dirname, "../ui/react/dist/kbd.js"),
-      "@docubook/ui-react/modal": path.resolve(__dirname, "../ui/react/dist/modal.js"),
-      "@docubook/ui-react/dropdown": path.resolve(__dirname, "../ui/react/dist/dropdown.js"),
-      "@docubook/ui-react/toggle": path.resolve(__dirname, "../ui/react/dist/toggle.js"),
-      "@docubook/ui-react/pagination": path.resolve(__dirname, "../ui/react/dist/pagination.js"),
+      "@docubook/ui-react/input": path.resolve(__dirname, "../ui-react/dist/input.js"),
+      "@docubook/ui-react/kbd": path.resolve(__dirname, "../ui-react/dist/kbd.js"),
+      "@docubook/ui-react/modal": path.resolve(__dirname, "../ui-react/dist/modal.js"),
+      "@docubook/ui-react/dropdown": path.resolve(__dirname, "../ui-react/dist/dropdown.js"),
+      "@docubook/ui-react/toggle": path.resolve(__dirname, "../ui-react/dist/toggle.js"),
+      "@docubook/ui-react/pagination": path.resolve(__dirname, "../ui-react/dist/pagination.js"),
       "@docubook/ui-react/theme-controller": path.resolve(
         __dirname,
-        "../ui/react/dist/theme-controller.js"
+        "../ui-react/dist/theme-controller.js"
       ),
-      "@docubook/ui-react/breadcrumbs": path.resolve(__dirname, "../ui/react/dist/breadcrumbs.js"),
-      "@docubook/ui-react/collapse": path.resolve(__dirname, "../ui/react/dist/collapse.js"),
-      "@docubook/ui-react/drawer": path.resolve(__dirname, "../ui/react/dist/drawer.js"),
-      "@docubook/ui-react/navbar": path.resolve(__dirname, "../ui/react/dist/navbar.js"),
+      "@docubook/ui-react/breadcrumbs": path.resolve(__dirname, "../ui-react/dist/breadcrumbs.js"),
+      "@docubook/ui-react/collapse": path.resolve(__dirname, "../ui-react/dist/collapse.js"),
+      "@docubook/ui-react/drawer": path.resolve(__dirname, "../ui-react/dist/drawer.js"),
+      "@docubook/ui-react/navbar": path.resolve(__dirname, "../ui-react/dist/navbar.js"),
     },
   },
 });